bug: correctly mark better benchmark in all cases
If two results differ by less than a nanosecond the wrong one is
occasionally marked as best.

The unwrap here should be okay since it is believed that Criterion
never stores NaN values. With that said, this probably should be fixed
to either remove NaN values or return an error or something similar,
since we don't have any guarantees to this effect. But this is fine for
now I think.

PR #13